apps/knowledge-builder/pyproject.toml
[project]
name = "knowledge-builder"
version = "0.1.0"
description = "Knowledge Builder for UIT AI Assistant - Crawl, process, and index documents"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "llama-index-core>=0.11.0",
    "llama-index-embeddings-openai>=0.1.0",
    "llama-index-llms-openai>=0.1.0",
    "llama-index-llms-gemini>=0.1.0",
    "llama-index-vector-stores-chroma>=0.1.0",
    "llama-parse>=0.5.0",
    "chromadb>=0.5.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "tiktoken>=0.5.0",
    "crawl4ai>=0.1.0",
    "playwright>=1.40.0",
    "aiohttp>=3.9.0",
    "llama-index-llms-ollama>=0.9.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.5",
    "pymupdf>=1.26.6",
    "pytesseract>=0.3.13",
    "pandas>=2.3.3",
    "streamlit>=1.30.0",
]

[project.scripts]
ukb = "main:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
